+++ python/tests/test_convert.py
@@ -86,4 +86,11 @@ print(s)
 print(".ip2int()",s.ip2int())
 print(nvstrings.int2ip(s.ip2int()))
 
+s = nvstrings.to_device(["2019-03-20T12:34:56.789Z","2020-02-29"])
+print(s)
+print(".timestamp2int()",s.timestamp2int())
+print(".timestamp2int(ms)",s.timestamp2int(units='milliseconds'))
+print(".int2timestamp()",nvstrings.int2timestamp([1553085296,1582934400]))
+print(".int2timestamp(ms)",nvstrings.int2timestamp([1553085296789,1582934400000],units='milliseconds'))
+
 s = None